requirements.txt
# Core dependencies
meraki==2.0.1
requests==2.32.3
python-dotenv==1.1.0

# Scheduling (optional - use APScheduler for scheduled runs)
APScheduler==3.10.4

# Development
pytest==7.4.4
